Python/SLA/requirements.txt
Flask
matplotlib==3.8.2
mysql-connector-python == 8.3.0
numpy==1.26.3
pandas==2.2.0
prometheus_api_client==0.5.4
scikit_learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
statsmodels==0.14.1
